src/requirements.txt
# TBWA Agency Databank - Python Dependencies
# Core requirements for migrations, seeding, and automation

# Database
psycopg2-binary==2.9.9
alembic==1.13.1
sqlalchemy==2.0.25

# Environment
python-dotenv==1.0.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# CLI and utilities
click==8.1.7
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==22.0.0

# Code quality
black==23.12.1
isort==5.13.2
ruff==0.1.9
mypy==1.8.0

# API testing
requests==2.31.0
httpx==0.26.0

# n8n workflow testing
n8n-client==0.1.0  # If available, otherwise skip

# OCR (optional)
paddleocr==2.7.0  # For PaddleOCR integration
paddlepaddle==2.6.0  # CPU version

# Cloud providers
boto3==1.34.26  # AWS/S3
azure-storage-blob==12.19.0  # Azure
google-cloud-storage==2.14.0  # GCP

# Monitoring
sentry-sdk==1.39.2
prometheus-client==0.19.0

# Async support
asyncio==3.4.3
aiohttp==3.9.1
asyncpg==0.29.0

# Data processing
pandas==2.1.4
numpy==1.26.3

# Date/time
python-dateutil==2.8.2
pytz==2023.3

# Encryption
cryptography==41.0.7
bcrypt==4.1.2

# JWT
pyjwt==2.8.0

# YAML/TOML parsing
pyyaml==6.0.1
toml==0.10.2

# Logging
structlog==24.1.0
python-json-logger==2.0.7

# File handling
openpyxl==3.1.2  # Excel
python-magic==0.4.27  # File type detection
Pillow==10.2.0  # Image processing

# Email
sendgrid==6.11.0
mailgun==0.1.1

# Scheduling
apscheduler==3.10.4

# Rate limiting
slowapi==0.1.9

# Progress bars
tqdm==4.66.1

# Makefile helper (for Python-based tasks)
invoke==2.2.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Type hints
types-requests==2.31.0.20231231
types-pytz==2023.3.1.1